remove dead types
